refactor: moved SpecFromElem to spec_from_elem.rs
